move postprocess and helper panel computations to helper function file
